update NPB kernel
