+++ micro-benchmarks/DRB175-barrier-orig-no.c
@@ -1,3 +1,11 @@
+/*
+!!!~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~!!!
+!!! Copyright (c) 2017-20, Lawrence Livermore National Security, LLC
+!!! and DataRaceBench project contributors. See the DataRaceBench/COPYRIGHT file for details.
+!!!
+!!! SPDX-License-Identifier: (BSD-3-Clause)
+!!!~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~!!!
+ */
 /**
  * IS.C: This file is part of kernel of the NAS Parallel Benchmarks 3.0 IS suit.
  * Intel Inspector can not correctly analysis the master, and report a false postive.
@@ -12,15 +20,16 @@ int main(int argh, char* argv[]){
   
   #pragma omp parallel private(i)
     for( i=0; i<10; i++){
-       #pragma omp master /*intel*/
-        {
-            q[i] = i;
-            for (j=0; j<10;j++)
-                qq[j] = q[j];
-        }
-        #pragma omp barrier /*Tsan*/
-          for (j=0; j<10;j++) qq[j] = 0;
-        #pragma omp for nowait
-          for (j=0; j<10;j++) qqq[j] = q[j];
+      #pragma omp master /*intel*/
+      {
+        q[i] = i;
+        for (j=0; j<10;j++)
+        qq[j] = q[j];
+      }
+      #pragma omp barrier /*Tsan*/
+        for (j=0; j<10;j++) qq[j] = 0;
+      #pragma omp for nowait
+        for (j=0; j<10;j++) qqq[j] = q[j];
     }
+    return 0;
 }
