+++ micro-benchmarks/DRB176-master-orig-no.c
@@ -1,3 +1,11 @@
+/*
+!!!~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~!!!
+!!! Copyright (c) 2017-20, Lawrence Livermore National Security, LLC
+!!! and DataRaceBench project contributors. See the DataRaceBench/COPYRIGHT file for details.
+!!!
+!!! SPDX-License-Identifier: (BSD-3-Clause)
+!!!~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~!!!
+ */
 /**
  * SP.C: This file is part of kernel of the NAS Parallel Benchmarks 3.0 SP suit.
  * Intel Inspector can not correctly analysis the master, and report a false postive.
@@ -12,13 +20,14 @@ int main(int argh, char* argv[]){
   
   #pragma omp parallel private(i)
     for( i=0; i<10; i++){
-       #pragma omp master /*intel*/
+      #pragma omp master /*intel*/
         {
             q[i] = i;
             for (j=0; j<10;j++)
                 qq[j] = q[i];
         }
         #pragma omp barrier /*Tsan*/
-        for (j=0; j<10;j++) qq[j] = 0;
+          for (j=0; j<10;j++) qq[j] = 0;
     }
+   return 0;
 }
